Move utils to centralised folder
