+++ clients/ui/frontend/src/__tests__/cypress/cypress/tests/mocked/modelCatalog/modelCatalog.cy.ts
@@ -16,7 +16,7 @@ import {
 } from '~/__mocks__';
 import { MODEL_CATALOG_API_VERSION } from '~/__tests__/cypress/cypress/support/commands/api';
 import { mockCatalogFilterOptionsList } from '~/__mocks__/mockCatalogFilterOptionsList';
-import { SourceLabel, type CatalogSource } from '~/app/modelCatalogTypes';
+import { SourceLabel, type CatalogSource } from '~/app/shared/types/catalogTypes';
 import { ModelRegistryMetadataType } from '~/app/types';
 import { ModelCatalogStringFilterKey } from '~/concepts/modelCatalog/const';
