+++ clients/ui/frontend/src/app/api/modelCatalog/service.ts
@@ -1,14 +1,16 @@
 import { APIOptions, handleRestFailures, isModArchResponse, restGET } from 'mod-arch-core';
 import {
-  CatalogArtifactList,
-  CatalogFilterOptionsList,
   CatalogLabelList,
-  CatalogModel,
-  CatalogModelList,
-  CatalogPerformanceArtifactList,
   CatalogSourceList,
   CatalogSourceListParams,
   CatalogLabelListParams,
+} from '~/app/shared/types/catalogTypes';
+import {
+  CatalogArtifactList,
+  CatalogFilterOptionsList,
+  CatalogModel,
+  CatalogModelList,
+  CatalogPerformanceArtifactList,
   ModelCatalogFilterStates,
   PerformanceArtifactsParams,
 } from '~/app/modelCatalogTypes';
