+++ clients/ui/frontend/src/app/context/mcpCatalog/types.ts
