+++ clients/ui/frontend/src/app/context/modelCatalogSettings/ModelCatalogSettingsContext.tsx
@@ -4,7 +4,8 @@ import useModelCatalogSettingsAPIState, {
   ModelCatalogSettingsAPIState,
 } from '~/app/hooks/modelCatalogSettings/useModelCatalogSettingsAPIState';
 import { useCatalogSourceConfigs } from '~/app/hooks/modelCatalogSettings/useCatalogSourceConfigs';
-import { CatalogSourceConfigList, CatalogSourceList } from '~/app/modelCatalogTypes';
+import type { CatalogSourceList } from '~/app/shared/types/catalogTypes';
+import type { CatalogSourceConfigList } from '~/app/modelCatalogTypes';
 import { BFF_API_VERSION, URL_PREFIX } from '~/app/utilities/const';
 import useModelCatalogAPIState from '~/app/hooks/modelCatalog/useModelCatalogAPIState';
 import { useCatalogSourcesWithPolling } from '~/app/hooks/modelCatalogSettings/useCatalogSourcesWithPolling';
