+++ clients/ui/frontend/src/app/hooks/modelCatalog/useModelCatalogAPI.ts
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { ModelCatalogContext } from '~/app/context/modelCatalog/ModelCatalogContext';
 import { ModelCatalogAPIState } from './useModelCatalogAPIState';
