+++ clients/ui/frontend/src/app/pages/mcpCatalog/screens/__tests__/McpCatalogGalleryView.spec.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import { McpCatalogContext } from '~/app/context/mcpCatalog/McpCatalogContext';
-import type { McpCatalogContextType } from '~/app/pages/mcpCatalog/types/mcpCatalogContext';
+import type { McpCatalogContextType } from '~/app/context/mcpCatalog/types';
 import type { McpServer } from '~/app/mcpServerCatalogTypes';
 import type { McpServersResult } from '~/app/hooks/mcpServerCatalog/useMcpServersBySourceLabel';
 import { useMcpServersBySourceLabelWithAPI } from '~/app/hooks/mcpServerCatalog/useMcpServersBySourceLabel';
