+++ clients/ui/frontend/src/app/pages/modelCatalog/components/ModelCatalogNumberFilter.tsx
@@ -2,7 +2,7 @@ import { Content, ContentVariants, FormGroup, NumberInput } from '@patternfly/re
 import * as React from 'react';
 import { ModelCatalogNumberFilterKey } from '~/concepts/modelCatalog/const';
 import { useCatalogNumberFilterState } from '~/app/pages/modelCatalog/hooks/useCatalogFilterState';
-import { CatalogFilterNumberOption } from '~/app/modelCatalogTypes';
+import { CatalogFilterNumberOption } from '~/app/shared/components/catalog';
 
 type ModelCatalogNumberFilterProps<K extends ModelCatalogNumberFilterKey> = {
   title: string;
