+++ clients/ui/frontend/src/app/shared/components/catalog/CatalogSourceLabelToggle.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 import { ToggleGroup, ToggleGroupItem } from '@patternfly/react-core';
-import type { CatalogLabelList, CatalogSourceList } from '~/app/modelCatalogTypes';
-import { SourceLabel } from '~/app/modelCatalogTypes';
+import type { CatalogLabelList, CatalogSourceList } from '~/app/shared/types/catalogTypes';
+import { SourceLabel } from '~/app/shared/types/catalogTypes';
 import {
   filterEnabledCatalogSources,
   getLabelDisplayName,
   getUniqueSourceLabels,
   hasSourcesWithoutLabels,
   orderLabelsByPriority,
-} from '~/app/pages/modelCatalog/utils/modelCatalogUtils';
+} from './utils/catalogSourceUtils';
 
 type SourceLabelBlock = {
   id: string;
